requirements.txt
lxml==4.2.1
pyodbc==4.0.23
python-docx==0.8.6
